feat: refactor py model device
