+++ rtp_llm/models_py/modules/factory/attention/__init__.py
@@ -1,137 +1,19 @@
-"""Attention factory module - handles different attention implementations."""
+"""Attention factory module - handles different attention implementations.
 
-import logging
+Attention impl selection is driven by:
+  1. Registry: AttentionImplEnum (qualname strings, no imports)
+  2. Device: get_prefill_mha_priorities() etc. (returns Enum list)
+  3. Selector: attn_factory.get_fmha_impl() (lazy load + support() check)
+"""
 
-# Import the factory after lists are defined to avoid circular imports
 from rtp_llm.models_py.modules.factory.attention.attn_factory import AttnImplFactory
 from rtp_llm.models_py.modules.factory.attention.fmha_impl_base import (
     FMHAImplBase,
     MlaImplBase,
 )
-from rtp_llm.ops.compute_ops import DeviceType, get_exec_ctx
 
 __all__ = [
     "FMHAImplBase",
     "MlaImplBase",
     "AttnImplFactory",
 ]
-
-# ============================================================================
-# Device-specific Attention implementation registration
-# ============================================================================
-from rtp_llm.models_py.modules.factory.attention.attn_factory import (
-    DECODE_MHA_IMPS,
-    DECODE_MLA_IMPS,
-    PREFILL_MHA_IMPS,
-    PREFILL_MLA_IMPS,
-)
-
-device_type = get_exec_ctx().get_device_type()
-if device_type == DeviceType.ROCm:
-    # Import to register ROCm FMHA implementations
-    from rtp_llm.models_py.modules.factory.attention.rocm_impl.aiter import (
-        AiterDecodeImplAsm,
-        AiterDecodeImplNonAsm,
-        AiterDecodeImplTriton,
-        AiterPrefillImplAsm,
-        AiterPrefillImplNonAsm,
-        AiterPrefillImplPaged,
-    )
-
-    PREFILL_MHA_IMPS.append(AiterPrefillImplPaged)
-    PREFILL_MHA_IMPS.append(AiterPrefillImplAsm)
-    PREFILL_MHA_IMPS.append(AiterPrefillImplNonAsm)
-    DECODE_MHA_IMPS.append(AiterDecodeImplAsm)
-    DECODE_MHA_IMPS.append(AiterDecodeImplNonAsm)
-    DECODE_MHA_IMPS.append(AiterDecodeImplTriton)
-else:
-    # currently append early means impl has higher priority
-    if device_type == DeviceType.Cuda:
-        from rtp_llm.models_py.modules.factory.attention.cuda_headwise_impl.headwise import (
-            HeadWisePrefillImpl,
-        )
-        from rtp_llm.models_py.modules.factory.attention.cuda_headwise_impl.headwise_fp8 import (
-            HeadWiseFP8PrefillImpl,
-        )
-        from rtp_llm.models_py.modules.factory.attention.cuda_impl.py_flashinfer_mha import (
-            PyFlashinferDecodeImpl,
-            PyFlashinferPagedPrefillImpl,
-            PyFlashinferPrefillImpl,
-        )
-        from rtp_llm.models_py.modules.factory.attention.cuda_impl.trt import (
-            TRTMHAImpl,
-            TRTPagedMHAImpl,
-        )
-        from rtp_llm.models_py.modules.factory.attention.cuda_impl.trtllm_gen import (
-            FlashInferTRTLLMDecodeImpl,
-            FlashInferTRTLLMPrefillImpl,
-            FlashInferTRTLLMSpecDecodeImpl,
-        )
-        from rtp_llm.models_py.modules.factory.attention.cuda_impl.xqa import (
-            get_xqa_impl,
-        )
-
-        PREFILL_MHA_IMPS.extend(
-            [
-                HeadWiseFP8PrefillImpl,
-                HeadWisePrefillImpl,
-                FlashInferTRTLLMSpecDecodeImpl,
-                FlashInferTRTLLMPrefillImpl,
-                TRTMHAImpl,
-                PyFlashinferPrefillImpl,
-                PyFlashinferPagedPrefillImpl,
-                TRTPagedMHAImpl,
-            ]
-        )
-        DECODE_MHA_IMPS.extend([FlashInferTRTLLMDecodeImpl])
-        DECODE_MHA_IMPS.append(get_xqa_impl())
-
-        from rtp_llm.models_py.modules.factory.attention.cuda_mla_impl.flashinfer_mla_wrapper import (
-            MlaFlashInferDecodeImpl,
-            MlaFlashInferPrefillImpl,
-        )
-
-        DECODE_MLA_IMPS.append(MlaFlashInferDecodeImpl)
-        PREFILL_MLA_IMPS.append(MlaFlashInferPrefillImpl)
-
-        # SparseMlaImpl requires CUDA >= 12.9 for flash_mla support
-        try:
-            import torch
-
-            if torch.version.cuda:
-                major, minor = map(int, torch.version.cuda.split(".")[:2])
-                if (major, minor) >= (12, 9):
-                    from rtp_llm.models_py.modules.factory.attention.cuda_mla_impl.flashmla_sparse_cp_impl import (
-                        SparseMlaCpImpl,
-                    )
-                    from rtp_llm.models_py.modules.factory.attention.cuda_mla_impl.flashmla_sparse_impl import (
-                        SparseMlaImpl,
-                    )
-
-                    DECODE_MLA_IMPS.append(SparseMlaImpl)
-                    PREFILL_MLA_IMPS.append(SparseMlaImpl)
-                    PREFILL_MLA_IMPS.append(SparseMlaCpImpl)
-        except (ImportError, AttributeError, ValueError):
-            pass  # Skip SparseMlaImpl if CUDA < 12.9 or flash_mla not available
-
-        from rtp_llm.models_py.modules.factory.attention.cuda_impl.flash_infer import (
-            FlashInferDecodeImpl,
-            FlashInferPrefillImpl,
-        )
-
-        PREFILL_MHA_IMPS.append(FlashInferPrefillImpl)
-        DECODE_MHA_IMPS.append(FlashInferDecodeImpl)
-
-    from rtp_llm.models_py.modules.factory.attention.cuda_impl.py_flashinfer_mha import (
-        PyFlashinferDecodeImpl,
-        PyFlashinferPrefillImpl,
-    )
-
-    PREFILL_MHA_IMPS.append(PyFlashinferPrefillImpl)
-    DECODE_MHA_IMPS.append(PyFlashinferDecodeImpl)
-
-    from rtp_llm.models_py.modules.factory.attention.cuda_cp_impl.prefill_cp_flashinfer import (
-        CPFlashInferImpl,
-    )
-
-    PREFILL_MHA_IMPS.append(CPFlashInferImpl)
