+++ rtp_llm/models_py/modules/factory/attention/cuda_impl/trtllm_gen.py
@@ -537,6 +537,10 @@ class FlashInferTRTLLMDecodeOp(object):
 
 class FlashInferTRTLLMPrefillImpl(FMHAImplBase):
 
+    @classmethod
+    def is_available(cls, fmha_config):
+        return fmha_config is None or not fmha_config.disable_flash_infer
+
     def __init__(
         self,
         attn_configs: AttentionConfigs,
@@ -601,6 +605,10 @@ class FlashInferTRTLLMPrefillImpl(FMHAImplBase):
 
 
 class FlashInferTRTLLMSpecDecodeImpl(FMHAImplBase):
+
+    @classmethod
+    def is_available(cls, fmha_config):
+        return fmha_config is None or not fmha_config.disable_flash_infer
 
     def __init__(
         self,
@@ -677,6 +685,10 @@ class FlashInferTRTLLMSpecDecodeImpl(FMHAImplBase):
 
 
 class FlashInferTRTLLMDecodeImpl(FMHAImplBase):
+
+    @classmethod
+    def is_available(cls, fmha_config):
+        return fmha_config is None or not fmha_config.disable_flash_infer
 
     def __init__(
         self,
