+++ rtp_llm/models_py/modules/factory/fused_moe/__init__.py
@@ -1,6 +1,7 @@
 """FusedMoe factory module
 
 Uses strategy pattern and builder pattern for refactored MOE factory.
+Device-specific strategy registration is driven by device.get_moe_strategy_candidates().
 
 Main components:
 - FusedMoeFactory: Main factory class
@@ -16,10 +17,9 @@ Usage example:
     moe = FusedMoeFactory.create_fused_moe(config, weights)
 """
 
-import torch
+import logging
 
-from rtp_llm.models_py.utils.arch import get_sm, is_cuda
-from rtp_llm.ops.compute_ops import DeviceType, get_exec_ctx
+from rtp_llm.device import get_current_device
 
 from .defs.fused_moe import FusedMoe
 from .factory import FusedMoeFactory
@@ -31,75 +31,12 @@ __all__ = ["FusedMoeFactory", "StrategyRegistry", "FusedMoe"]
 # Device-specific MoE strategy registration
 # ============================================================================
 
-device_type = get_exec_ctx().get_device_type()
+registry = StrategyRegistry()
 
-# Import common strategies
-from rtp_llm.models_py.modules.factory.fused_moe.impl.common.strategy.batched_triton_strategy import (
-    BatchedTritonStrategy,
-)
+for strategy_cls in get_current_device().get_moe_strategy_candidates():
+    try:
+        registry.register(strategy_cls())
+    except Exception as e:
+        logging.debug(f"Skipping MoE strategy {strategy_cls.__name__}: {e}")
 
-if device_type == DeviceType.ROCm:
-    # ========== ROCm Registry ==========
-
-    # MoE strategies
-    from rtp_llm.models_py.modules.factory.fused_moe.impl.rocm.strategy import (
-        RocmBf16PureTPStrategy,
-        RocmEpLowLatencyStrategy,
-        RocmEpNormalStrategy,
-        RocmFp8PerChannelPureTPStrategy,
-    )
-    registry = StrategyRegistry()
-    registry.register(RocmEpLowLatencyStrategy())
-    registry.register(RocmEpNormalStrategy())
-    registry.register(RocmFp8PerChannelPureTPStrategy())
-    registry.register(RocmBf16PureTPStrategy())
-    registry.register(BatchedTritonStrategy())
-    FusedMoeFactory.set_registry(registry)
-
-else:
-    # ========== CUDA Registry ==========
-
-    # MoE strategies
-    from rtp_llm.models_py.modules.factory.fused_moe.impl.cuda.strategy import (
-        CudaFp8PerBlockEpLowLatencyStrategy,
-        CudaFp8PerBlockEpNormalStrategy,
-        CudaFp8PerBlockNoDPMaskedStrategy,
-        CudaFp8PerBlockNoDPStrategy,
-        CudaFp8PerTensorEpLowLatencyStrategy,
-        CudaFp8PerTensorEpNormalStrategy,
-        CudaFp8PerTensorNoDPStrategy,
-        CudaNoQuantCppStrategy,
-        CudaNoQuantDpNormalStrategy,
-        CudaNoQuantEpLowLatencyStrategy,
-        CudaW4a8Int4PerChannelEpLowLatencyStrategy,
-        CudaW4a8Int4PerChannelEpNormalStrategy,
-        CudaW4a8Int4PerChannelNoDPStrategy,
-    )
-
-    registry = StrategyRegistry()
-    registry.register(CudaFp8PerTensorEpLowLatencyStrategy())
-    registry.register(CudaFp8PerTensorEpNormalStrategy())
-    registry.register(CudaFp8PerBlockEpLowLatencyStrategy())
-    registry.register(CudaFp8PerBlockEpNormalStrategy())
-    registry.register(CudaFp8PerBlockNoDPMaskedStrategy())
-    registry.register(CudaFp8PerBlockNoDPStrategy())
-    registry.register(CudaFp8PerTensorNoDPStrategy())
-    registry.register(CudaNoQuantEpLowLatencyStrategy())
-    registry.register(CudaNoQuantDpNormalStrategy())
-    registry.register(CudaNoQuantCppStrategy())
-    registry.register(BatchedTritonStrategy())
-    registry.register(CudaW4a8Int4PerChannelEpLowLatencyStrategy())
-    registry.register(CudaW4a8Int4PerChannelEpNormalStrategy())
-    registry.register(CudaW4a8Int4PerChannelNoDPStrategy())
-    # Only register FP4 strategies on SM_100+ (and only if CUDA GPU is available)
-    if torch.cuda.is_available() and is_cuda() and get_sm()[0] >= 10:
-        from rtp_llm.models_py.modules.factory.fused_moe.impl.cuda.strategy import (
-            CudaFp4EpLowLatencyStrategy,
-            CudaFp4EpNormalStrategy,
-            CudaFp4NoDPStrategy,
-        )
-
-        registry.register(CudaFp4EpLowLatencyStrategy())
-        registry.register(CudaFp4EpNormalStrategy())
-        registry.register(CudaFp4NoDPStrategy())
-    FusedMoeFactory.set_registry(registry)
+FusedMoeFactory.set_registry(registry)
