+++ rtp_llm/models_py/modules/hybrid/causal_attention.py
@@ -4,18 +4,12 @@ import torch
 import torch.nn as nn
 
 from rtp_llm.models_py.distributed.collective_torch import Group, all_reduce
+from rtp_llm.models_py.modules.base import FusedQKRMSNorm
 from rtp_llm.models_py.modules.factory import LinearFactory
 from rtp_llm.models_py.modules.factory.attention.fmha_impl_base import FMHAImplBase
 from rtp_llm.ops import AttentionConfigs, HWKernelConfig, ParallelismConfig
-from rtp_llm.ops.compute_ops import DeviceType, LayerKVCache, get_exec_ctx
+from rtp_llm.ops.compute_ops import LayerKVCache
 from rtp_llm.utils.model_weight import W
-
-# Import device-specific FusedQKRMSNorm
-device_type = get_exec_ctx().get_device_type()
-if device_type == DeviceType.ROCm:
-    from rtp_llm.models_py.modules.base.rocm.norm import FusedQKRMSNorm
-else:
-    from rtp_llm.models_py.modules.base.cuda.norm import FusedQKRMSNorm
 
 
 class CausalAttention(nn.Module):
