+++ rtp_llm/models_py/utils/arch.py
@@ -2,23 +2,17 @@ from typing import Tuple
 
 import torch
 
-from rtp_llm.ops.compute_ops import DeviceType, get_exec_ctx
-
 
 def is_cuda():
-    device_type = get_exec_ctx().get_device_type()
-    if device_type == DeviceType.Cuda:
-        return True
-    else:
-        return False
+    from rtp_llm.device import get_current_device
+
+    return get_current_device().is_cuda()
 
 
 def is_hip():
-    device_type = get_exec_ctx().get_device_type()
-    if device_type == DeviceType.ROCm:
-        return True
-    else:
-        return False
+    from rtp_llm.device import get_current_device
+
+    return get_current_device().is_rocm()
 
 
 def get_num_device_sms() -> int:
